[ONNXifi] Add support for ReplaceNaN operator
Description:
This commit adds support for the `ReplaceNaN` operator to the ONNX model
importer. Most of the work to support this operator was done in an
earlier commit that added support for the Caffe2 loader. Consequently,
this commit only includes code that parses inputs, outputs, and arguments
from an .onnxtxt and calls `Graph::createReplaceNaN` to implement the
operation.

Testing:
This commit adds a unit test to `onnxImporterTest.cpp` to test loading
this operator from a .onnxtxt file.