[ONNXifi] Add support for BatchBoxCox operator
Description:
This commit adds support for the BatchBoxCox operator to the ONNX model
importer. Most of the work to support this operator was done in an
earlier commit that added support to the Caffe2 loader. Consequently,
this commit only includes code that parses inputs, outputs, and arguments
from an .onnxtxt and calls Graph::createBatchBoxCox to implement the
operation.

Testing:
This commit adds a unit test to onnxImporterTest.cpp to test loading
this operator from a .onnxtxt file.