[Caffe2] Add support for SparseToDense operator
Description:
This commit adds support for the `SparseToDense` Caffe2 operator. This
operator takes a sparse representation of a tensor specified by an array
of indices and an array of tensor values and converts it into a dense
representation. In this dense representation, values[i] is stored at
indices[i], and the rest of the dense representation is zero-filled.
Indices can appear twice in the sparse representation, and in this case, all
corresponding values are summed and stored in dense representation.

This operator is implemented as its own Glow node and instruction.

Testing:
This commit adds tests for this operator to `caffe2ImporterTest.cpp` and
`OperatorTest.cpp`.