[Convert] Teach convertOutputs how to deal with module level value
*Description*
This patch teaches the convertOutputs method how to deal with module
level value. In particular, such values have one conversion per function
instead of one for the whole conversion process.
Moreover, when they are used as addresses in save nodes, the conversion
must happen on the input value not the output address.
E.g., when converting
save origTy input, origTy toVal

The correct result is not:
save origTy input, convertto newTy toVal to origTy

But:
save convertto origTy input to newTy, newTy toVal

With this commit, this change is NFC, because the functionality of
changing module level values is not exposed.

*Testing*
None, NFC