[Image] Fix writePngImage 90 degree rotation and black image output issues
*Description*
The 90 degree rotation was simply a swapping of the column and row indices
when accessing the tensor.
The black image output was a misuse of the range input. The range represents
the range of values the image should have, not the range of values the input
tensor has.

In other words, we were doing:
- read: tensor = image * scale + bias
- write: image = tensor * scale + bias

Whereas we should be doing:
- read: tensor = image * scale + bias
- write: image = (tensor - bias ) / scale

Fixing both these issues makes the image being written properly.

Additionally, there was a bug in the way we released the memory.
We were using png_destroy_read_struct instead of png_destroy_write_struct
and that could result in segfaults depending on the libpng version.

*Testing*
The added test loads an existing image into a tensor, then writes it
into an image and finally load that image back.
The tensor obtained from the final image loading is expected to be
the same (modulo scale and bias rounding) as the one from the
original image.