[ONNX] Add ONNX support for SparseToDense operator
Description:
This commit adds support for importing the `SparseToDense` operator from
an ONNX model and running it on the CPU backend.

Testing:
This commit modifies the existing operator test for `SparseToDense` to
run on both CPU and interpreter, and adds a test to the ONNX importer
test to make sure it is imported correctly from a .onnxtxt file.