[Caffe2/ONNX] Add support for LengthsSum operator
Description:
This commit adds support for the `LengthsSum` operator, which
performs a `BatchedReduceAdd` operation on groups of consecutive
zero-axis slices of the `data` input. The sizes of these groups
are defined by the `lengths` input. This operator has implementations in
both the CPU and interpreter backends, and can be imported from a Caffe2
and ONNX model.

Testing:
This commit adds an interpreter and CPU operator test as well as Caffe2
and ONNX importer tests.