[Interpreter] Add FP16 support for the full family of inference instruction
*Description*

The interpreter supported the instructions needed to run resnet50 in FP16.
With this patch, we refactor the code such that now all the instructions
used in inference support both fp32 and fp16.
The refactoring consists in templatizing the based implementation
and adds a dispatch method to use the appropriate fp16 or fp32 type.

*Testing*
This commit does not add any tests because the implementation uses the
same code as the fp32 one and that should be covered in the existing tests.
The dispatch mechanism is also already tested by previously added
fp16 tests.