Run a round of DCE before any other graph optimizations
Optimizer may be called after backend specific transformations and some nodes may have become unused. It is a good idea to remove them, before proceeding with any further optimizations. Without it some optimizations do not kick in as they may check for the number of users of a given node and count the unused nodes as users as well.