[Verifier] Switch the node verifier from assert to returning bool for the nodes
*Description*
The verifier is a valuable tool to validate if an input graph is valid.
Right now, a lot of checks are *not* done in the loader because it knows
the verifier will catch them. However, the verifier does nothing
in release mode and thus we cannot rely on it in that mode.

To be able to use the verifier in release mode, we need to refactor it
so that it does not rely on asserts anymore.

This patch goes in that direction by doing the switch for the classes
related to Node.
Moreover, since the use case from the verifier switches from purely
debugging to error reporting, this patch improves the way error are
reported. In particular, it prints details on the node that fails so
that a potential user doesn't have to debug the compiler to see what
is wrong with the input graph.

E.g., before this patch, the graphTest would print:
Assertion failed: (*getType() == getPayload().getType() &&
                   "Underlying tensor type doesn't match constant type"),
function verify, file ../lib/Graph/Nodes.cpp, line 298.

Now we would get:
In 'input'
Underlying tensor type doesn't match constant type
For comparison `LHS Equal RHS` with:
LHS: float16<5>
RHS: float<5>

That means we don't need to run the debugger anymore to see what is
wrong and we get this diagnostic in release mode as well.
The bonus point is that we can now test the verifier as we add more tests.
(This patch doesn't add all those tests.)

Couple of design points:
- All the reporting goes through glow::report so that if we want a mode
  where all the reports are silented, we just have to change this function
- The intermediate call to verify in the compiler will stay but would be
  wrapped into assert so that we don't pay the verify price in release mode

The next step is to switch the graph and module, then to add a verifier
call in the loader that would trigger proper error reporting (instead of
crashing).

*Testing*
Updated existing tests.
Ran resnet50 in release and debug mode.

Related to issue #1517