[FunctionQuantizer] Fix rescaling happening in the middle of nodes
*Description*
Some nodes, at least in their interpreter implementation, expect that
their inputs and outputs are scaled with the exact same parameters.
These constraints were not enforced in the function quantizer for
the slice, transpose, and reshape nodes leading to a mismatch between
how the consumers of a value expect it to be quantized and how it
was actually quantized.

This patch teaches the function quantizer how to fix those constraints
for these additional nodes (the mechanism already existed) and adds
checks in the verifier to catch this kind of problem earlier.

*Testing*
Added test cases, but only for two nodes. Given they all use the same
code path I didn't bother adding them all.

Related to issue #1989.