[examples] Add cmake for ResNet50 bundle
Description:
This commit adds the capability to build the ResNet50 standalone bundle
using `cmake` in addition to `make`. There are separate targets for
quantized and non-quantized version of the same bundle so that both can
be built in the same build tree.

Testing:
Configured fresh build and ran bundle running cmake targets.