[Verifier] Switch the node verifier from assert to returning bool for the function and module
*Description*
This patch extends the work already done for the nodes to the function
and module. Essentially, this moves the verify method of the related
classes from assert based to returning a bool.
Thus, the caller can decided whether this should only be run in assert
mode or anytime, e.g., for error checking.

On top of that, this commit provides more context when things break
using the same framework (VerifierHelper) that we introduced for the
nodes.

*Testing*
Updated test cases

Related to #1517.