[examples] Add cmake for LeNet MNIST bundle
Description:
This commit adds the capability to build the LeNet MNIST standalone bundle
using `cmake` in addition to `make`. There are separate targets for
quantized and non-quantized version of the same bundle so that both can
be built in the same build tree.

Testing:
Configured fresh build and ran bundle running cmake targets.