[Nodes] Make lengths input of `BatchOneHot` int32
Description:
This commit modifies the `BatchOneHot` operator to use 32-bit integer
values in the `lengths` input rather than 64-bit integers. This change
was supposed to be part of an earlier commit that migrated the index
inputs of all sparse operators to 32-bit integer, but it was overlooked
at the time.

Testing:
This commit modifies the operator test for `BatchOneHot` to use 32-bit
integers. The modified test passes, as do all other unit tests.