[tests] Leverage LIT and FileCheck infrastructure when available
*Description*
This patch adds a couple of tests based on the LIT and FileCheck
tools from LLVM for the text-translator.
When both lit.py and FileCheck are found in the path, a new set
of tests, litTests, are added to the list of available tests.

This gives a convenient way to test our command line tools. This
new tests, only kicks in when glow is built in release mode, since
otherwise, they would be too slow.

*Testing*
Added lit.py and FileCheck in the path
Ran cmake -GNinja -DCMAKE_BUILD_TYPE=Release -DGLOW_MODELS_DIR=<downloadedModelsPath>
Ran ninja check