[utils] Add a convenient script to set the llvm testing environment
*Description*
This patch adds a script that will add lit.py and FileCheck to the
PATH if they are not already present.
To use it one simply need to source it.

This script downloads llvm and build what is needed to use LIT.

*Testing*
source utils/set_llvm_test_env.sh
cmake <...>

Then checked that HAS_LLVM_TEST_ENV is ON