[Runtime] Add runtime Executor class
Description:
This commit adds `Executor`, an interface for a component of the Glow
runtime that is capable of running a partitioned directed acyclic Glow
graph and returning the results of that execution. This commit also adds
`ThreadPoolExecutor`, an implementation of this interface that runs
graphs using a thread pool whose workers usher every received request to
completion through a series of state transitions.

Testing:
It compiles.

Documentation:
Comments.