[Runtime] Add runtime Executor class
Description:
This commit adds `Executor`, an interface for a component of the Glow
runtime that is capable of running a partitioned directed acyclic Glow
graph and returning the results of that execution. This commit also adds
`ThreadPoolExecutor`, an implementation of this interface that runs
graphs using a thread pool. The  workers essentially do a
concurrent breadth-first search through the node graph to execute all
nodes, repeatedly copying partition components outputs to their descendents along
the way.

Testing:
This commit adds eight unit tests for the following scenarios:
1) empty graph
2) single node
3) several instances of a single node graph at once
4) duplicate runIds
5) multi node graph
6) multi node graph with a node that fails
7) multi node graph with nodes running on multiple devices
8) several instances of a multi node graph at once

Documentation:
Comments.