[examples][cifar10] Add VGG like model, some refactoring
*Description*:
Add a 2nd model to the cifar10 example that is close to VGG as implemented in pytorch vision vgg.py / vgg11. This is currently missing the `Dropout` layers as I couldn't find a builtin node for that in glow.

*Testing*:
Ran the cifar10 example with `-model-vgg`, accuracy is improving more steadily than with the jumpy behavior of the default model.

*Documentation*:
Related to #2111