[NodeBuilder] Add Input and Output Indices enums (#2207)
As mentioned in https://github.com/pytorch/glow/issues/2194#issuecomment-449008434, there are places where we use indices explicitly to specify inputs or outputs of Nodes. It's a really bad practice as these indices could change if the Node definition or NodeBuilder change. This PR improves the situation by adding input and output indices enums to each node.

There's a commit here with a couple examples of places it can be used to clean things up. There are other places to update if we agree this approach makes sense. It could also be useful for use in the new design of `isOpSupported()` (https://github.com/pytorch/glow/issues/2194).

Here's what the enums look like for the `ConvolutionNode` in `AutoGenNodes.h` with this PR:

```cpp
class ConvolutionNode final : public Node {
  ...
  enum InputIndices {
    InputIdx = 0,
    FilterIdx = 1,
    BiasIdx = 2,
  };

  enum ResultIndices {
    ResultIdx = 0,
  };
```