[C2/ONNX] Add support for GatherRanges operator
Description:
This commit adds support for the `GatherRanges` operator. This operator
is not lowered and this commit includes interpreter and CPU backend
implementations for it.

Testing:
This commit adds two unit tests (one with 32-bit ranges and another with
64-bit ranges) to `OperatorTest` that check operator correctness and one
unit test to `OnnxImporterTest` to check that the operator can be loaded
from a `.onnxtxt` correctly.