[Runtime] Fix heap-use-after-free TSAN warning in ThreadPoolExecutor
Description:
This commit fixes a heap-use-after-free bug in `ThreadPoolExecutor`
detected by TSAN. The symptom of the bug is that on some invocations of
`ExecutorTest`, the `ThreadPoolExecutor` gets destroyed before
`ThreadPoolExecutor::handleDeviceManagerResult` finishes executing. This
happens because the callback provided in `ThreadPoolExecutor::run()` is
called *before* erasing the execution state for the run that just
finished. The invocation of the callback fills the `std::future` that the
test thread waits on before calling EXPECT* and ASSERT* on the values
returned by the `ThreadPoolExecutor`. Sometimes, the test thread checks
the return values and exists the scope of the test before the rest of
`handleDeviceManagerResult` can execute, which causes
`executionStateMap_` to be destroyed before the thread executing
`handleDeviceManagerResult` calls `erase` on it.

This commit fixes this bug by adding an `InflightBarrier` class  that
keeps track of calls to `DeviceManager::runFunction` that the executor
hasn't received callbacks for or has not finished processing the result of.
The barrier is incremented before calling `DeviceManager::runFunction`
and decremented after. `~ThreadPoolExecutor` waits for this barrier to
hit zero before allowed the destruction of the object.

Testing:
Ran ExecutorTest with TSAN 100 times locally and with ASAN 50 times
locally.