[InstGen] LocalResponseNormalization src/dest should not be in-place
Description:
This commit removes code that marks the source and destination buffers
of the LocalResponseNormalization instruction as in-place. The value of
an output element depends on the values adjacent to the corresponding
input element, so there is no way to perform the local response
normalization operation in-place.

Testing:
All unit tests pass.