[Operator] Add ElementwiseLinear Caffe2 op
Description:
This commit adds support for the `ElementwiseLinear` Caffe2 operator. It
is lowered early in the `glow::Function::createElementwiseLinear`
implementation to `Broadcast`, `Mul` and `Add`.

Testing:
This commit adds a test for this operator to `OperatorTest` (CPU,
Interpreter and OpenCL backends) and two tests to `Caffe2ImporterTest`;
one that has `axis` specified in the test protobuf and one that does
not. These new tests and all old unit tests all pass.