[unittests] Enable each test case for specific backends in OperatorTest.cpp. (#2476)
*Description*:

What we'd need is to specific backends that each test case is enabled for, however, GTEST is incapable of parameterizing each test case separately since GTEST's parameterization only works per harness.
We could create a new harness for a test case to achieve that, but it's too verbose. This diff proposes a compromise:
We put all test cases under a single harness instantiated for all possible backends -- as a result all test cases will be running for all backends. To whitelist specific backends for each specific test case, ENABLED_BACKENDS(backend1, backend2, ...) is used as a guard for each test to enable it only for specific backends.

*Testing*:

Ran OperationTest to see it passes.

*Documentation*:

Fixes #2170