[Training] Add support for Adagrad training algorithm
Description:
This commit adds support for the Adagrad (adaptive gradient) training
algorithm. This algorithm computes a unique learning rate for each
trainable model parameters that is directly proportional to the sum of
the squares of past gradients with respect to that weight. Weights with
large gradients are paired with small learning rates, and vice versa.

Testing:
This commit adds a test case that learns an exponential function using
the Adagrad algorithm.