[OpenCL] Account for a case were device max WG size limits dims.
In OpenCL, the device can define maximum sizes separately for each
dimension, but the total maximum WG size can be less than the
multiple of all of the dimension maximums. In that case it simply
means that not all dimensions can be set at their maximum at the
same time. This patch accounts for that and reduces the second
dimension to meet the maximum WG size limit of the device.