[Tests] Split RuntimeUtils and add ExecutorBenchmarkWrapper
Description:
This commit splits up the utility class inside RuntimeUtils so that
ExecutorTest is not forced to include benchmarking utilities and adds
ExecutorBenchmarkWrapper, a subclass of ExecutorTestWrapper that accepts
a benchmarking state object as a parameter. This is used to pause timing
while executing code that sets up for execution and checks the results for
correctness.

Testing:
All unit tests pass.