[Benchmarks] Add runtime benchmark framework
Description:
This commit adds a Glow runtime benchmark that runs the same model using
various components of the runtime stack in order to determine how much
overhead each one contributes to the overall latency.

The benchmark defines a template class named RuntimeBenchmark that extends
benchmark::Fixture. The template parameter is meant to be a specific
backend class (e.g. CPUBackend). This class provides a virtual runBenchmark() method for
subclasses to override in order to run the benchmarks they represent and
some common code to prepare Backend, Module and ExecutionContext
instances for use in the benchmarks.

There is a subclass of RuntimeBenchmark for each component of the inference runtime stack.
Each subclass maintains members for additional state it requires (e.g.
list of DeviceManagers in the case of ExecutorBenchmark) as well as code
to set up/tear down that state. That code sets everything up correctly
and needs to be overridden to benchmark special cases like device
failure.

There are two macros that are used to create the actual benchmarks.
DECLARE_RUNTIME_BENCHMARK(name, moduleCreatorFn, dagCreatorFn) declares
subclasses of HostManagerBenchmark, ExecutorBenchmark and
DeviceManagerBenchmark whose setUpModule and setUpDAG methods are
overriden using moduleCreateFn and dagCreatorFn.
INSTANTIATE_RUNTIME_BENCHMARK(name, backend) instantiates all three
kinds of benchmarks for the given backend. moduleCreatorFn and dagCreatorFn
are user-defined functions that create a Module and Executor DAG for
benchmarking, respectively.

Testing:
All unit tests pass and new benchmarks run successfully (except for
Interpreter backend).