[habana] Enable pipelining and multi-device support for HabanaDeviceManager
Summary:
This commit enables pipelining and multi-device execution support for
HabanaDeviceManager. There are now two thread pools in `HabanaDeviceManager`, one
that executes `runFunction` requests and passes the wait handles to the second
pool that waits on them and then calls the provided callback. Invocations of
Synapse API functions that are not thread-safe or need to be atomic are guarded
by process-wide locks.

Reviewed By: bertrand

Differential Revision: D14552718