[Executor] Fix placeholder propagation bug
Description:
This commit fixes a bug in which intermediate outputs generated by partitions
are not propagated to their children. This is because the executor
assumes that DeviceManager::runFunction will return a set of PlaceholderBindings
containing the output Placeholders, but this is not true.
DeviceManager::runFunction needs to be invoked with PlaceholderBindings
that contain Tensors for both input and output Placeholders.

This commit fixes this bug by creating Placeholders for all intermediate
DAG node outputs, storing them in the ExecutionState corresponding
to the run that needs them, and mapping them to zero Tensors in the
input bindings for the DAG node that produces them.

The executor unit test has been modified accordingly.

Testing:
All unit tests pass.