[runtime] Use default eviction callback in HostManager
Summary: Passing nullptr here is kind of brittle; it assumes that the backend's
DeviceManager is aware of the contract that "nullptr" here means "provide a
default behavior".  Instead, let's just provide a NOP callback.  The DM is
always free to log (or not) as it sees fit.

Test Plan: A bit hard to test, but if you remove the default behavior from
CPUDeviceManager and run resnet-runtime, you'll get a runtime failure.