build: only support LLVM 8
Adjust the build to use the `REQUIRED` parameter to `find_package` as we want to
drop support for LLVM 6, and LLVM 8 is the first version available built with
the C++11 ABI.

Make the llvm-link search more resilient to the future by using the
LLVM_VERSION_MAJOR variable to add the suffix.  Additionally, when building
against the build tree, fetch the exported target for llvm-link and use that.