[tensor] Improve file I/O routine
Description: Needing to know the tensor type out-of-band makes it a lot harder
to use these routines for instrumentation and debugging.  Since these tensor
dumps aren't intended to be a real serialization format, we can just
fwrite/fread the Type struct to get the information we want.

Test Plan: Modify existing test to not rely on type knowledge.