[OpenCL] Add OpenCL implementation of SparseLengthsWeightedSum
Description:
This commit adds an OpenCL implementation of SparseLengthsWeightedSum.
It is very similar to the CPU backend implementation; the only difference
is that the outputs for each segment are computed in parallel.

Testing:
This commit enables SparseLengthsSum and SparseLengthsWeightedSum operator tests for
OpenCL and they pass. All other unit tests pass.