[OpenCL] Add OpenCL implementation of SparseLengthsWeightedSumGrad
Description:
This commit adds an OpenCL implementation of
SparseLengthsWeightedSumGrad. This implementation is identical to the
CPU backend implementation. It is difficult to parallelize this
operation without giving each global ID its own copy of the dataGrad
buffer and adding together the copies at the end. As a result, this has
been left as a future work item.

Testing:
This commit enables learnSparseLengthsWeightedSum for the OpenCL backend in
MLTest and it passes.