[ir-optimizer] Add option -instrument-debug-only to instrument only instructions from the provided list
Only instructions whose names contain a substring from the list will be debug-instrumented. This is very useful for debugging big NN models, where dumping inputs and outputs of all instructions usually produces huge logs, which are hard to navigate. Limiting the printing to specific instructions significantly reduces the size of the logs and shows only what the user wants to see.