[googlebenchmark] Update googlebenchmark submodule
Description:
This commit updates the googlebenchmark submodule inside the tests
directory. This update also includes an explicit setting of CMP0063
(cmake policy related to dynamic linking visibilty of symbols in a target that is linked
statically and then dynamically) that eliminates the cmake warning
related to some targets in that submodule when configuring the Glow
project.

Testing:
Ran cmake.