Avoid copying onnxifi weights where possible (#2921)
Summary:
Make it so that Constants can have unowned tensors as their payload and will basically copy-on-write by making sure that when a mutable reference to the payload is requested, the payload will be copied to a new owned tensor if it was previously unowned.
The motivation behind this is to load weights from onnxifi using unowned tensors as often as possible only making a copy if the weight needs to be modified by Glow.
Because Glow currently lacks a UInt8QTy quantized type, all tensors of this type must still be copied and converted to Int8QTy but if a UInt8QTy type is added to Glow then these can be taken as unowned tensors as well.
One thing to consider here is that the lifetime of the memory backing Constants created at model loading time must live for at least as long as the Constants. For our current model loading this is true but it is something that will have to be maintained.

Documentation:
Doxygen
Pull Request resolved: https://github.com/pytorch/glow/pull/2921

Differential Revision: D15354037

Pulled By: jackm321

fbshipit-source-id: 1e74b89f7f8f69331844c83c70812c5a191c8553