[IROpt] Improve copy elimination optimisation
Description:
This commit improves the copy instruction elimination optimisation in
the case that the src and dest operands are tensorview instructions.
Prior to this commit, this optimisation used to eliminate copy
instructions if the src an dest tensorview offsets were equal. This does
not work for all cases; for example, offsets {0} and {0, 0} are not
programatically equal, but are semantically equal.

This commit modifies this optimisation to check that the linearized
offsets into the underlying buffers after peeling away all tensorviews
are equal before erasing the copy instruction. For example, the previous
implementation incorrectly eliminated a copy between identical offsets
into two different intermiediate tensorviews that provide a view into
different sections of the same buffer. This commit addresses this.

This commit also modifies verify for Insert/ExtractTensorInst and
TensorViewInst to check that the number of offsets in the instruction is equal to the size of the
src/dest shape. This simple check precludes the case in which offsets {0} and {0,
0} are compared by this optimization and determined not to be equal.
Code that creates these types of instruction has also been updated
accordingly.

Testing:
This commit adds a test to IROptTest that check that copies to and from
tensorviews with identical linearized offsets into the same buffer are detected and
optimized out, another test to check that the copy is not optimized out
in the previous case if the linearized offsets are not equal, and a
test to check that a copy from a tensorview to its source buffer are
detected and optimized out.

All unit tests pass.