Add helper function to dump a DAG after partitioning. (#2936)
Summary:
This PR dump a dot file from the partition result. Since now all functions belong to a function family and they have the same partition, we only dump the one function's partition. For each node, so far we show the function name and the backendKind, which are the most important info in my mind.
The following is the Heterogeneous partition example:
![dag1](https://user-images.githubusercontent.com/37386895/57913545-aa7c7000-7841-11e9-86b7-4fd1c6593fd7.png)
Documentation:
Pull Request resolved: https://github.com/pytorch/glow/pull/2936

Differential Revision: D15434311

Pulled By: beicy

fbshipit-source-id: 804ede444abe71cd1acbd12f9a9893546e0fa248