Change command line memory options to kilobytes (#2956)
Summary:
CI is broken because the `en2gr_cpu_partition_test` tries to set the CPU memory on the command line, but it seems that command line parameters don't work with `uint64_t` on some platforms. I've hit this issue in the past, and I'm not sure why it's the case.
Pull Request resolved: https://github.com/pytorch/glow/pull/2956

Reviewed By: bertmaher

Differential Revision: D15458152

Pulled By: jfix71

fbshipit-source-id: de91dd71ba7182d54fd7219f52091f009f51d720