Bug fix: remove duplicated generated "save" node in heterogeneous partition. (#2965)
Summary:
In Heterogeneous partition, after backendBasedPartition, we got a list of function. Then when we finally create the DAG, it is possible that one input in one function may use the output of another function. In that case, we don't need to generate an "save" node again. i

Documentation:
Pull Request resolved: https://github.com/pytorch/glow/pull/2965

Differential Revision: D15485353

Pulled By: beicy

fbshipit-source-id: c3ed56b53c00aa749ac6fdcd70887396c8204a6a