Disable buggy opt foldChannelShuffle
Summary:
This optimization to replace reshape->transpose->reshape with
ChannelShuffle breaks patterns where the output type isn't the same as the
input type.  I think the pattern check is not strict enough; it seems like it
needs to check the output type as well as which indices are swizzled by the
transpose to make sure it's doing the right thing.

GitHub Test Plan: New test that demonstrates fold() -- which calls
foldChannelShuffle -- breaking (if it is reenabled).

Differential Revision: D15508242

fbshipit-source-id: 317cee6af7d075e4f5761cd94b0ab9239ff37e1d