Disable buggy opt foldChannelShuffle (#2983)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/2983

This optimization to replace reshape->transpose->reshape with
ChannelShuffle breaks patterns where the output type isn't the same as the
input type.  I think the pattern check is not strict enough; it seems like it
needs to check the output type as well as which indices are swizzled by the
transpose to make sure it's doing the right thing.

GitHub Test Plan: New test that demonstrates fold() -- which calls
foldChannelShuffle -- breaking (if it is reenabled).

Differential Revision: D15508242

fbshipit-source-id: 3e4a3412997aec04e1f1e47146085e6802fab484