Prevent integer overflow when using uniform_int_distribution (#3000)
Summary:
In the recomandationSystemTest we were generating random data with a range of -INT32_MAX - INT32_MAX, which caused an integer overflow inside uniform_int_distribution. It seems internally in uniform_int_distribution there is a max-min operation which when using the full range will cause an overflow.  The fix is to make the distribution of type long long which will allow the full int32 range.

Documentation: NA
Pull Request resolved: https://github.com/pytorch/glow/pull/3000

Differential Revision: D15546825

Pulled By: gcatron

fbshipit-source-id: 270d2eb30104740ba530dd7dd5d3f199772294b1