build: try to goad CMake into understanding dependencies
Change the build rules for the JIT to use some more trickery to have CMake track
dependencies itself.  First, we swap out the compiler around the definition of
JIT module to ensure that we build with Clang.  Second, we build the library as
a module to control the link rule invocation.  Third, we change the linker
language to LLIR and provide a custom linker rule to invoke the LLVM linker.
This should hopefully provide more reliable null builds.  As a happy side
effect, the logic is easier to read.