Improve RecSys so that float MLP produces results other than 0. (#3020)
Summary:
the last FC layer of top-level MLP is essentially a dot product of non-negative vector with a lot of 0 by a vector with equal number of positives and negatives. Due to high number of 0s in LHS vector, it is very likely that we'll get a negative number after reduce, which turns into 0 because of the following Relu. I used skewed distribution instead, so that RHS vector has more positive values, and result is more likely to be positive.
Pull Request resolved: https://github.com/pytorch/glow/pull/3020

Differential Revision: D15573174

Pulled By: rdzhabarov

fbshipit-source-id: d991b90107419810de7b43dd4d6049c2fe36c390