Add support for Complete events (#2995)
Summary:
In the Chromium TraceEvent spec, Complete events encode an entire TraceEvent using begin timestamp and duration. In many cases this allows us to replace two events (one for the begin, one for the end) with a single event with an additional parameter. This diff adds support for Complete events in the Glow TraceEvent framework and converts as many of our trace logging events as possible to use this new format.

Changes involved:
* Modified the ScopedTraceBlock to log Complete events at the exit of scope, rather than a Begin at in the constructor and an End at the exit.
* Added new macros for easier usage of scoped trace event blocks: a definition to early exit the scope (we use generally before firing callbacks), and versions which provide an event variable name, which allows nesting scoped events.
* Converted almost all TraceEvents in the runtime to use the SCOPE macros (and so Complete events).
* Updated auto instrumentation on Interpreter, CPU and OpenCL to use Complete events rather than a Begin and End pair. Manual TraceEvents which specify Begin and End still work. This actually simplifies the instrumentation code a bit.
* Changed Interpreter, CPU and OpenCL devices to use std::system_clock timestamp domain to match the Runtime and Habana device. This required a little fudging of device timestamps for OCL which we should come back to.

There should be no functional change to viewing traces, but trace file size is as much as 45% reduced. (verified across resnet-runtime and tracing-compare examples).

Documentation: No changes, we still support the TraceEvent spec.
Pull Request resolved: https://github.com/pytorch/glow/pull/2995

Differential Revision: D15577954

Pulled By: nickgg

fbshipit-source-id: 72157f73071734a10b37e87e18ff82a0212b5ff4